requirements.txt
gradio
google-generativeai
selenium
webdriver-manager
beautifulsoup4
trafilatura
yfinance
python-dotenv